1st security

+++ src/main/java/web/service/UserService.java
@@ -1,5 +1,6 @@
 package web.service;
 
+import org.springframework.security.core.userdetails.UserDetails;
 import web.model.User;
 
 import java.util.List;
@@ -7,9 +8,10 @@ import java.util.List;
 public interface UserService {
     public void saveUser(User user);
 
-    public void deleteUser(int id);
+    public void deleteUser(Long id);
 
-    public User getUserByID(int id);
+    public User getUserByID(Long id);
 
     public List<User> listUsers();
+
 }
